Fix target weight helper imports
